extract utils functions from appStore
