+++ src/stores/appStore.ts
@@ -3,10 +3,11 @@ import { createStore } from 'solid-js/store';
 import {
   type Status,
   type Destination,
-  type Path,
   VPNService,
 } from '../services/vpnService';
 import { AppStoreContext } from './appContext';
+import { buildLogContent } from '../utils/status';
+import { areDestinationsEqualUnordered } from '../utils/destinations';
 
 export type AppScreen = 'main' | 'settings' | 'logs';
 
@@ -45,79 +46,12 @@ export function createAppStore() {
     response?: import('../services/vpnService').StatusResponse;
     error?: string;
   }) => {
-    let content: string | undefined;
-    if (args.response) {
-      const statusValue = args.response.status;
-      if (typeof statusValue === 'object') {
-        if ('Connected' in statusValue || 'Connecting' in statusValue) {
-          const isConnected = 'Connected' in statusValue;
-          const destination = (statusValue as any)[
-            isConnected ? 'Connected' : 'Connecting'
-          ] as import('../services/vpnService').Destination;
-          const city = destination.meta?.city || '';
-          const location = destination.meta?.location || '';
-          const where = [city, location].filter(Boolean).join(', ');
-          content = `${isConnected ? 'Connected' : 'Connecting'}: ${where} - ${
-            destination.address
-          }`;
-        }
-      } else if (statusValue === 'Disconnected') {
-        const lastWasDisconnected =
-          state.logs.length > 0 &&
-          state.logs[state.logs.length - 1].message.startsWith('Disconnected');
-        if (lastWasDisconnected) {
-          content = undefined;
-        } else {
-          const lines = args.response.available_destinations.map(d => {
-            const city = d.meta?.city || '';
-            const location = d.meta?.location || '';
-            const where = [city, location].filter(Boolean).join(', ');
-            return `- ${where} - ${d.address}`;
-          });
-          content = `Disconnected. Available:\n${lines.join('\n')}`;
-        }
-      } else {
-        const statusLabel = statusValue;
-        const destinations = args.response.available_destinations.length;
-        content = `status: ${statusLabel}, destinations: ${destinations}`;
-      }
-    } else if (args.error) {
-      content = `${args.error}`;
-    }
+    const lastMessage = state.logs.length
+      ? state.logs[state.logs.length - 1].message
+      : undefined;
+    const content = buildLogContent(args, lastMessage);
     if (!content) return;
     appendContentIfNew(content);
-  };
-
-  const canonicalizeMeta = (
-    meta: Record<string, string> | undefined
-  ): string => {
-    if (!meta) return '';
-    const keys = Object.keys(meta).sort();
-    const ordered: Record<string, string> = {};
-    for (const key of keys) ordered[key] = meta[key];
-    return JSON.stringify(ordered);
-  };
-
-  const canonicalizePath = (path: Path): string => {
-    if ('Hops' in path) return `Hops:${path.Hops}`;
-    return `IntermediatePath:${(path.IntermediatePath || []).join(',')}`;
-  };
-
-  const areDestinationsEqualUnordered = (
-    a: Destination[],
-    b: Destination[]
-  ): boolean => {
-    if (a.length !== b.length) return false;
-    const aSorted = [...a].sort((x, y) => x.address.localeCompare(y.address));
-    const bSorted = [...b].sort((x, y) => x.address.localeCompare(y.address));
-    for (let i = 0; i < aSorted.length; i += 1) {
-      const da = aSorted[i];
-      const db = bSorted[i];
-      if (da.address !== db.address) return false;
-      if (canonicalizeMeta(da.meta) !== canonicalizeMeta(db.meta)) return false;
-      if (canonicalizePath(da.path) !== canonicalizePath(db.path)) return false;
-    }
-    return true;
   };
 
   const getStatus = async () => {
